Finish making sure the space is available
- Worked with my classmates a bit and were able to modify my code in
events.js so that I am actually referencing the correct data.  I was
pretty close before but we weren't calling the data correctly so
that the computer will know that the space has something there.
- We created a gameArray variable that is equal to store.game.cells to
correctly reference our game index array.  And based on the box selected
the command will either be sent through to the API request (if the space
is empty and thus complete the turn by placing the marker in the space)
or it will indicate that the space is already taken and prompt
the player to try again.