Finish Second Round UI Steps
- Created a function in events.js that will automatically alternate
between the first player ('X') and second player ('O') after their mark
has been placed on the board successfully.
- Added JQuery script to update an h3 in index.html to allow the
user to know when the turn has changed.
- Created a new file gamewins.js to host my function which will allow
the players to know when one of them has won the game by getting 3 of
their mark in a row (vertical/horizontal/diagonal) or that the game is a
draw.