Complete Update Game portion of project
- Wrote additional parts to previous if else statements in events.js
click function so that when the game is decided to be a win, the API
request will provide the game.over a variable of true. and finish the
game.
- I added a .off() function to the event listener in my onGameOver
success function in ui.js to stop players from being ablet to click on
the board once the game is over.
- Moved my original event listener to the onNewGame function in my
events.js file so that when you click the new game button - the board
will once again be clickable.